+++ cli/src/commands/file/mod.rs
@@ -14,8 +14,11 @@
 
 mod annotate;
 mod chmod;
+mod delete;
+mod edit;
 mod list;
 mod search;
+mod set;
 mod show;
 mod track;
 mod untrack;
@@ -29,8 +32,11 @@ use crate::ui::Ui;
 pub enum FileCommand {
     Annotate(annotate::FileAnnotateArgs),
     Chmod(chmod::FileChmodArgs),
+    Delete(delete::FileDeleteArgs),
+    Edit(edit::FileEditArgs),
     List(list::FileListArgs),
     Search(search::FileSearchArgs),
+    Set(set::FileSetArgs),
     Show(show::FileShowArgs),
     Track(track::FileTrackArgs),
     Untrack(untrack::FileUntrackArgs),
@@ -44,8 +50,11 @@ pub async fn cmd_file(
     match subcommand {
         FileCommand::Annotate(args) => annotate::cmd_file_annotate(ui, command, args).await,
         FileCommand::Chmod(args) => chmod::cmd_file_chmod(ui, command, args).await,
+        FileCommand::Delete(args) => delete::cmd_file_delete(ui, command, args).await,
+        FileCommand::Edit(args) => edit::cmd_file_edit(ui, command, args).await,
         FileCommand::List(args) => list::cmd_file_list(ui, command, args).await,
         FileCommand::Search(args) => search::cmd_file_search(ui, command, args).await,
+        FileCommand::Set(args) => set::cmd_file_set(ui, command, args).await,
         FileCommand::Show(args) => show::cmd_file_show(ui, command, args).await,
         FileCommand::Track(args) => track::cmd_file_track(ui, command, args).await,
         FileCommand::Untrack(args) => untrack::cmd_file_untrack(ui, command, args).await,